+++ src/shell_executor_lib/__init__.py
@@ -1,3 +1,3 @@
 """Exposed shell_executor_lib classes and methods."""
-from shell_executor_lib.errors import CommandError
+from shell_executor_lib.errors import CommandError, AuthenticationError, PrivilegesError
 from shell_executor_lib.manager.command_manager import CommandManager