+++ src/shell_executor_lib/errors/__init__.py
@@ -1,2 +1,4 @@
 """Exposed errors classes and methods."""
 from shell_executor_lib.errors.command_error import CommandError
+from shell_executor_lib.errors.authentication_error import AuthenticationError
+from shell_executor_lib.errors.privileges_error import PrivilegesError