+++ tests/mock_shell_executor_lib/__init__.py
@@ -1,4 +1,6 @@
 """Exposed mock classes and methods."""
 from mock_shell_executor_lib.mock_command_executor import (mock_subprocess_error_exit_code, mock_subprocess_correct,
                                                            mock_subprocess_invalid_exit_code,
-                                                           mock_subprocess_invalid_outs)
+                                                           mock_subprocess_invalid_outs, mock_output_data)
+
+mock_asyncio_subprocess = "asyncio.create_subprocess_shell"