+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - @javierfh03 - Fix error when execute complex commands.
 - @javierfh03 - Fix error when return additional line in list in the output from command.
+- @javierfh03 - Fix wait to execute command.
 
 ## [0.1.1] - 2024-02-14
 ### Changed