+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2024-02-17
+## [0.2.0] - 2024-02-29
+### Changed
+- @javierfh03 - Add Authentication and Privileges errors and changes in the command managers structure.
 ### Fixed
 - @javierfh03 - Fix error when execute complex commands.
 - @javierfh03 - Fix error when return additional line in list in the output from command.
@@ -19,5 +21,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - @javierfh03 - Add command manager with units test.
 
-[0.1.1]: https://github.com/Lagatrix/shell_executor_lib
+[0.2.0]: https://github.com/Lagatrix/shell_executor_lib
+[0.1.1]: https://github.com/Lagatrix/shell_executor_lib/releases/tag/0.1.1
 [0.1.0]: https://github.com/Lagatrix/shell_executor_lib/releases/tag/0.1.0